parallelization, body_headeR_line sample names

+++ output_file.py
@@ -91,8 +91,9 @@ class Output_file:
                     self.check_if_input_file_invalid()
 
         if self.invalid is True:
-            if os.path.isfile(self.path):
-                os.remove(self.path)
+            if self.path:
+                if os.path.isfile(self.path):
+                    os.remove(self.path)
             return False
         else:
             return True
@@ -107,9 +108,13 @@ class Output_file:
         self.version = self.list_of_input_files[0].version
 
     def extract_chromosomes(self):
-        for input_file in self.list_of_input_files:
-            input_file.extract_indices_for_chromosomes()
-            self.chromosomes_position.update(input_file.chromosomes_positions)
+        with ThreadPoolExecutor(max_workers=10) as executor:
+            [executor.submit(self.extract_indices_for_chrom_in_file, input_file) for input_file in
+             self.list_of_input_files]
+
+    def extract_indices_for_chrom_in_file(self, input_file):
+        input_file.extract_indices_for_chromosomes()
+        self.chromosomes_position.update(input_file.chromosomes_positions)
 
     def process_headers(self):
         """ Delete duplicates and sort all lists regarding the header. Creates a body header line according to the
@@ -133,10 +138,10 @@ class Output_file:
         list_of_chrom.sort(key=lambda x: self.alphanum_key(x))
 
         for chrom in list_of_chrom:
-            self.list_of_body_records_chrom.clear() 
+            self.list_of_body_records_chrom.clear()
             with ThreadPoolExecutor(max_workers=10) as executor:
                 [executor.submit(self.multithread_test,input_file,chrom) for input_file in self.list_of_input_files]
-                
+
             self.adjust_body_records_to_samples()
             self.list_of_body_records_chrom = list(toolz.unique(self.list_of_body_records_chrom, key=lambda x: x.line))
             self.list_of_body_records_chrom.sort(key=lambda x: self.alphanum_key(x.line))
@@ -231,6 +236,9 @@ class Output_file:
 
     def create_body_header_line_for_output(self):
         """ Creates a body header line according to the samples. """
+        if len(self.list_of_samples_to_be_combined) == 0:
+            self.determinate_samples_to_be_combined()
+        Body_header_line.list_of_samples_to_be_combined = self.list_of_samples_to_be_combined
         self.body_header_line = Body_header_line("")
         self.body_header_line.has_format_field = len(Body_header_line.list_of_samples_to_be_combined) > 0
         self.body_header_line.samples_names = Body_header_line.list_of_samples_to_be_combined
